requirements.txt
Click==7.0
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.6.0
numpy==1.24.2
pandas==1.5.3
opencv_python==4.7.0.72
Pillow==9.3.0
preprocessing==0.1.13
pyparsing==2.4.0
pytz==2020.1
scikit_learn==1.2.0
scipy==1.10.1
seaborn==0.9.0
six==1.12.0
torch==2.0.0
torchvision
